PR 19008: Fixes issue where warnings were being displayed in red
Fixes issue where warning displays in red because the `chalk` package (which gulp-core-build uses to colorize text) automatically detects whether stdout supports color. When running as a child process, this is false, and so the warning was never colorized to yellow. Rush then writes the entire contents of stderr in red after the build (unless, of course, the text has already been colorized). The correct solution then is to override the behavior of supports-color by adding a --color flag to the build command when Rush builds a project. This will have the added benefit that we will now see colors when running a build without the --quiet flag.