PR 31143: Rush should build dependencies along the critical path
Since we have limited parallelization, we should ensure that we are always building projects which are blocking the most other projects (e.g. things along the critical path - which is the limiting factor in build time).. This actually has a positive perf impact.. Builds have improved in speed by over 1 minutes on a few trials on my machine.

Related work items: #168346