PR 30774: Adding "publish" command to rush.
This adds an rush publish command to rush. It is responsible for consuming change files that look like:

```json
{
  "changes": [{
    "packageName": "@microsoft/rush-lib",
    "type": "major",
    "comment": "Patch change to rush-lib"
  }]
}
```

It processes one or more of these and and applies the changes intelligently:

* It bumps the version intelligently (highest order change wins)
* It bumps intra-repo dependency versions if necessary and re publishes patches of packages if their semver no longer meets the requirement.
* It deletes the processed change file(s)
* It git adds, commits, and pushes the changes
* It publishes the changed packages
* It tags the checkin appropriately.

This is the first initial checkin. Adds the basic workflow plus tests to validate it handles change requests correctly. It does not modify CHANGELOG.md files yet.