PR 33206: Adds "rush build" which supports incremental build.
For each project that rush builds, when it succeeds, a `package-deps.json` file is dropped. This file contains all the git hashes for everything in the folder (recursive) that is not gitignored.

When a build runs again, it compares the current hashes with whats in the file, if the file exists. If they are the same, the build is skipped. If they are different, we build the project, and then ensure that all downstream dependencies are not skipped.

By default, the `rebuild` command is left unaffected. I've sublcassed it as the `build` command which supports incremental by default. It also supports a `--clean` flag which will skip the check and force a rebuild, which gives us a path towards removing the `rebuild` command if `build` is good enough.



- If you run `rush install` or `rush generate`, it may cause new packages to be downloaded which will not be caught by incremental build and could result in issues. We will fix this upstream to allow dependencies to be considered. (You can always fall back to `rush build --clean`.)

- If you run `rush build` and then run it again with `--production` or any new flag, it won't rebuild. (You can always fall back to `rush build --clean`.)

- Right now we are doing a shallow comparison of the hashes directly in rush code. Instead this will be moved upstream so that the comparison is consistently implemented.

Related work items: #214255