PR 33155: Fixing enforced API Extractor jsDoc errors
*** There is one addition to fixing API Extractor errors, it is a readonly check.
If there is a getter and no setter then that member is readonly.

This is the first step in the process of turning on API Extractor.
We have fixed the current jsDoc errors in projects where an API Extractor file exists in their config folder.
After this PR, and after webpack warnings are (fixed), we will make all jsDoc errors logWarnings to fail CI builds that don't follow enforced jsDoc rules.