PR 34643: Introduce a no-op option when Rush builds projects
Change the functionality of ProjectBuildTask to, if the `build` `test` or `clean` options are specified as ampty strings, that we should treat them as a no-op. This is useful in certain cases (such as the generator, which doesn't get build [although it should be written in TypeScript!]). It also unblocks the issue Marc was seeing.