Merged PR 52668: Rush should treat some dependencies as "implicitly pinned versions"
To mimic the behavior of NPM, and to reduce extra side-by-side versioning issues, Rush should treat all 1st level dependencies as a pinned version if we only consume 1 version in the repo.

Take the following packages and their dependencies as an example:

lib-a:
   * react@1.0.7
   * some-react-lib@1.0.0

lib-b:
  * react@1.0.7

some-react-lib:
  * react@>=1.0.0 <2.0.0

If you just normally ran `npm install` in `lib-a`, you would always get react@1.0.7, since the direct dependency would be installed first and would satisfy the range required by some-react-lib.

However, due to the `common/package.json` file, "lib-a" is treated as a first-order dependency, but "react" is not. This can lead to cases where we would end up installing a very new version of React at the root, but installing the specific version actually required by the repo several times in the rush-subfolders.