Anything written to stderr should cause process to fail (#495)
If the `shouldWarningsFailBuild === true`, then we should capture anything written to `process.stderr`. If anything significant is written to `stderr`, we will fail the build.

We have had several cases where things are writing to stderr but are making it into the master build (and causing developers to get the "success with warnings" message). This change will make it impossible to check in any changes that will cause errant warnings to be written to the console (assumign