Fix `gulp serve` in https mode for Node 8 (#502)
It looks like Node8 automatically inserts the `http2` module as a built-in module. Earlier versions of `gulp-connect` don't work with http2. Also, the `gulp-connect` library automatically uses `http2` if it is present in the `node_modules` (in Node 6) and automatically uses it in Node 8.

This behavior makes it not easy to tell whether you are using http2 or not. Since this serve task is not meant to be used in production (rather just for local developer testing), we decided to disable http2 entirely.