[WIP] - Attempt to locate packages by package alias when using the npm: prefix
Hello, would love some help here. No idea if this is working or not TBH and would love some help/guidance, or if you want to take this over and land it I'd be happy with that, but I'm very happy to continue contributing to this. I currently don't know if this works as I haven't been able to figure out how to link this locally, or get the tests working.

This is an attempt to add support for finding packages when their semver range starts with `npm:`, which is a useful mechanism for aliasing or forking packages. See #1279 for more details.

Refs #1279