Update RushInstallManager.ts
When `npm install` is ran in a directory containing the npm-shrinkwrap file - devDependencies are still being installed - while they are not included in the shrinkwrap file.

This does not align with the documented description for `rush install`:

> The "rush install" command installs package dependencies for all your projects, based on the shrinkwrap file that is created/updated using "rush update"

This is being observed when using npm@6.x. And potentially aligns with the comment in this forum post:

https://npm.community/t/npm-i-changed-my-npm-shrinkwrap-package-lock-why/190

That in npm@6 `install` can lead to updates in the shrinkwrap file.