[api-documenter] Change modifier output order to match TypeScript syntax
Unlike other languages like Java, in TypeScript, the order of
modifiers is defined strictly. Currently, it looks like api-documenter
uses alphabetical order, which is not the correct TypeScript modifier
order. It seems the correct order is

(`public` | `protected` | `private`)? (`static`? | `abstract`?) `override`? `readonly`?

No member can be `static` and `abstract` at the same time, so the
order of these two does not matter. api-documenter currently only
uses `protected`, `static`, `abstract` (new in this PR) and `readonly`,
so `public` (default!), `private` (no API!) and `override` (maybe
interesting for navigation, but not as a modifier) do not apply.

Added a code comment about the order and moved `readonly` to the
last position so that the overall modifier output order now matches
TypeScript syntax.