[heft] Redesign watch mode
Initially the rework of Heft targeted a model in which file watching
was fully owned by Heft, and individual plugins would simply be informed
of changed files (with hashes) by Heft.

This design proved to be a significant challenge due to the performance
of having a file watcher monitoring the entire project folder while the
build was running. Specifically, when a large number of file changes
occurred in a short period of time, the watcher would emit a single
event indicating that an unknown amount of changes had occurred,
resulting in significant CPU and file system activity to identify what
change occurred in order to properly update the list of changed files.

In theory, this issue could be mitigated with access to the file system
change journal or other API that could provide a list of changes since
a previous marker, however there is also the issue that not all input
files to any given build step will be within the project folder. Tools
such as Webpack and TypeScript have logic to watch these files as well.

Thus, this change redesign's Heft's watch support to expect individual
plugins to manage their own file watchers, with a callback provided by
the Heft runtime that the plugin should invoke if it detects a relevant
change to its inputs, which will request Heft to invoke the plugin at
its earliest convenience.

When requested by a plugin, if Heft is currently executing a later task
in the build graph, it will cancel the build and start over, offering
every plugin a chance to perform new work as a result of file changes.
If the requesting plugin is still on the queue, the request is ignored.