[credential-cache] Support custom cache paths
The `CredentialCache` does not rely on the Rush user configuration. It can be used outside of `@rush/rush-sdk` by directly importing it from `@microsoft/rush-lib`.

To support such use cases, we should allow the consumer to specify a file path for their cache, rather than the fixed `~/.rush-user/credentials.json` path.

Non-breaking change because the default path remains the same, and we add support for additional options in `ICredentialCacheOptions`.

| `ICredentialCacheOptions.cacheDirectory` | `.cacheName` | Resulting Path |
|-|-|-|
| `undefined` | `undefined` | `~/.rush-user/credentials.json` (default) |
|  `undefined` | `custom-name` | `~/.rush-user/custom-name.json` |
| `/custom-dir` | `undefined` | `/custom-dir/credentials.json` |
| `/custom-dir` | `custom-name` | `/custom-dir/custom-name.json` |